Refractored and reindented code for mongo adapter, removed vcf adapter
