+++ scout/ext/backend/__init__.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 from __future__ import (print_function, absolute_import, unicode_literals)
 
-from .base import BaseAdapter
+from .event_handler import EventHandler
+from .variant_handler import VariantHandler
 from .fixture import FixtureAdapter
-from .vcf import VcfAdapter
 from .config_parser import ConfigParser
 from .mongo import MongoAdapter
 from .load_mongo import load_mongo_db
